requirements-base.txt
# Core dependencies
python-dotenv==1.0.0
pydantic-settings>=2.0.3
sqlalchemy==2.0.23
alembic==1.12.0
psycopg2-binary>=2.9.6
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.1
redis==4.6.0
prometheus-client==0.17.1
psutil==5.9.5